Handle IPv4-embedded IPv6 addresses with leading zeros
If the IPv6 address to be parsed ends with an IPv4 address in dot notation, the IPv4 portion will be represented by two parts in hexadecimal notation. If the first part is 0 (because the  IPv4 address in dot notation starts with "0.0."), it would fail to (correctly) parse the overall address, when the IPv6 prefix is not all zeros.
One practical use case of the dot notation is IPv4-embedded IPv6 addresses.  Without this change "::ffff:0.0.0.0" failed to parse, because it was rewritten to "::ffff:" and e.g. "::ffff:0.0.0.1" was rewritten to "::ffff:1", when it should have been "::ffff:0:1".

Leading zeros in the dot notation IPv4 part are now explicitly forbidden, because some systems interpret this as octal numbers. RFC 3986 defines "IPv4address" within an IPv6 address without leading zeros, so this decision has formal backing.